Rearrange js files. Fix logo size in about section
