+++ about.html
@@ -64,7 +64,7 @@
     </div>
     <div class="center-logo">
     <div class ="img-logo-container">
-      <a href="#"><img class="img-logo" src="styles-conference/assets/sd_logo.png" alt="img_logo"></a>
+      <a href="./index.html"><img class="img-logo" src="styles-conference/assets/sd_logo.png" alt="img_logo"></a>
     </div>
   </div>
   </section>
@@ -93,6 +93,5 @@
     </div> 
   </section>
 </body>
-<!-- <script type="module" src="main.js"></script>
-<script type="module" src="teachers-object.js"></script> -->
+<script type="module" src="main.js"></script>
 </html>
