+++ index.html
@@ -89,5 +89,6 @@
   </section>
 </body>
 <script type="module" src="main.js"></script>
-<script type="module" src="teachers-object.js"></script>
+<script type="module" src="dynamic_html_creator.js"></script>
+<!-- <script type="module" src="teachers-object.js"></script> -->
 </html>
