+++ styles-conference/stylesheets/styles.css
@@ -329,14 +329,14 @@ ul li img {
 .img-logo {
   width: 100%;
   height: 100%;
-  object-fit: cover;
+  object-fit: contain;
 }
 
 .img-logo-container {
   width: 80%;
   height: 120px;
   border: 1px solid #d3d3d3;
-  background-color: #f5f5f5;
+  background-color: white;
   overflow: hidden;
   margin-top: 25px;
 }
